+++ src/hooks/queries/useUserProfile.tsx
@@ -100,7 +100,7 @@ export const useUserProfile = (id: number): [UserProfile | null, any] => {
 			author: Number(id),
 			last: 5,
 		},
-		fetchPolicy: 'network-only',
+		// fetchPolicy: 'network-only',
 	});
 
 	// Prepare the credits data object.