+++ src/theme/global.ts
@@ -6,6 +6,13 @@ export const styles = {
 	global: {
 		body: { fontSize: '18px' },
 		textarea: { underline: 'none !important' },
+		p: {
+			marginTop: '0.5em',
+			marginBottom: '0.5em',
+		},
+		ul: {
+			listStylePosition: 'inside',
+		},
 
 		// WordPress content classes.
 		'.has-text-align-center': { textAlign: 'center' },