+++ src/views/LostPasswordView.tsx
@@ -1,14 +1,15 @@
-import { useCallback, useEffect, useState } from 'react';
+import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useGoogleReCaptcha } from 'react-google-recaptcha-v3';
 import { Button, Text, Flex, Container, Heading, Box, Spinner, useToast } from '@chakra-ui/react';
 
 import TextInput from '../components/common/inputs/TextInput';
 import { useSendPasswordResetEmail } from '../hooks/mutations/useSendPasswordResetEmail';
+import { useLostPasswordError } from '../hooks/hooks';
 
 export default function LoginView() {
 	const [username, setUsername] = useState<string>('');
-	const [reCaptchaToken, setReCaptchaToken] = useState<string>('');
+	const [errorCode, setErrorCode] = useState<string>('');
 	const { executeRecaptcha } = useGoogleReCaptcha();
 
 	const {
@@ -16,19 +17,15 @@ export default function LoginView() {
 		results: { loading: submitLoading },
 	} = useSendPasswordResetEmail();
 
-	const handleReCaptchaVerify = useCallback(async () => {
+	const handleReCaptchaVerify = async () => {
 		if (!executeRecaptcha) {
 			return;
 		}
 
-		const token = await executeRecaptcha('registerUser');
-		setReCaptchaToken(token);
-	}, [executeRecaptcha]);
+		const token = await executeRecaptcha('resetPassword');
 
-	// Trigger the verification as soon as the component is loaded
-	useEffect(() => {
-		handleReCaptchaVerify();
-	}, [handleReCaptchaVerify]);
+		return token;
+	};
 
 	const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
 		setUsername(e.target.value);
@@ -37,26 +34,33 @@ export default function LoginView() {
 	const toast = useToast();
 	const navigate = useNavigate();
 
-	// const errorMessage = useLoginError(errorCode);
+	const errorMessage = useLostPasswordError(errorCode);
 
 	const handleSubmit = (e: React.FormEvent) => {
 		e.preventDefault();
 
-		// TODO handle password reset errors
-		sendPasswordResetEmailMutation({ username, reCaptchaToken })
-			.then(() => {
-				toast({
-					title: 'Email sent',
-					description: 'Please check your inbox for reset instructions.',
-					status: 'success',
-					duration: 5000,
-					isClosable: true,
-					position: 'top',
+		handleReCaptchaVerify().then((token) => {
+			if (!token) {
+				setErrorCode('recaptcha_error');
+				return;
+			}
+
+			// TODO handle password reset errors
+			sendPasswordResetEmailMutation({ username, reCaptchaToken: token })
+				.then(() => {
+					toast({
+						title: 'Email sent',
+						description: 'Please check your inbox for reset instructions.',
+						status: 'success',
+						duration: 5000,
+						isClosable: true,
+						position: 'top',
+					});
+				})
+				.then(() => {
+					navigate('/');
 				});
-			})
-			.then(() => {
-				navigate('/');
-			});
+		});
 	};
 
 	return (
@@ -77,6 +81,7 @@ export default function LoginView() {
 						inputProps={{
 							autoComplete: 'username',
 						}}
+						error={errorMessage}
 					/>
 					<Flex justifyContent='flex-start' gap={4}>
 						<Button type='submit' colorScheme='blue' px={6}>